Refactor CommentController's save method to return ResponseEntity
This commit updates the save method in CommentController to return a ResponseEntity object. Previously, it returned a plain string indicating the request's success. Now, the method checks if the save operation is successful and, if so, fetches and returns a list of all comments associated with the board ID in an OK response. If the save operation fails, it returns a NOT_FOUND response indicating that the post does not exist.